Set project before batch processing
This enables more methods to be moved from QP to QPEx, and means scripts referring to projects can be run from the command line.
Restrict QP and QPEx to use ImageData<BufferedImage> (no need to be more generic now).
